Fix: Bug where job not being preselected and fix ui with badge
